requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==25.4.0
certifi==2024.8.30
coverage==7.13.4
Django==5.0.7
django-autocomplete-light==3.12.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djhacker==0.2.3
drf-spectacular==0.29.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
packaging==26.0
pillow==10.4.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.0.1
python-telegram-bot==21.5
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
sqlparse==0.5.1
uritemplate==4.2.0
